Update current.py
Added prettytable which allows for easy printing of tables and it gives headers for each line of output. If you think it looks messy I can keep looking around for solutions. To do this I had to change the line that compares fuzzy scores to convert both values to string but I checked and it looks like it still works. I added a while loop for the code so multiple sequences can be analyzed without having to execute the code multiple times.